changed /sensor/fluor msg type to Float32

+++ sampler/scripts/event_based_sampler_node.py
@@ -9,7 +9,7 @@ from bottle_state import Bottle
 
 from std_msgs.msg import Bool
 from std_msgs.msg import Int16
-from std_msgs.msg import Int32
+from std_msgs.msg import Float32
 from mavros_msgs.msg import RCIn
 from watersampling_msgs.msg import MS5837Stamped, PumpInfo
 
@@ -75,7 +75,7 @@ class SamplerNode():
         self.depth_sensor_sub = rp.Subscriber(
             '/watersampling/depth_sensor', MS5837Stamped, self.depthSensorCallback, queue_size=1)
         self.fluor_sensor_sub = rp.Subscriber(
-            '/sensor/fluor', Int32, self.fluorSensorCallback, queue_size=1)
+            '/sensor/fluor', Float32, self.fluorSensorCallback, queue_size=1)
 
         # ROS Publisher
         self.pump_info_pub = rp.Publisher(
